feat(layer2): Introduce shared types and traits for Layer 2 protocols
- Added `types.rs` to centralize shared types like `ProtocolState`, `AssetParams`, etc.
- Created `traits.rs` for defining common traits like `Proposal`, `ContractExecutor`, and `FederationMLHook`.
- Updated existing modules (RGB, RSK, Taproot Assets) to utilize the new shared types and traits.
- Enhanced RGB module with `RGBProposal` and `RGBManagerExt` for advanced flows.
- Improved RSK federation module with a robust implementation for cross-chain pegs.
- Refactored Taproot Assets module to support extensibility for contract execution and ML hooks.
- Added validation utilities in the security module for cryptographic operations and transaction inputs/outputs.
- Introduced `ValidationResult` struct for better validation feedback.
- Updated `Cargo.toml` for the new `anya-core` package structure and dependencies.

+++ anya-bitcoin/src/core/mempool/policy.rs
@@ -169,7 +169,7 @@ impl MempoolPolicy {
     
     /// Check if transaction size is within limits
     fn check_tx_size(&self, tx: &Transaction) -> Result<(), PolicyError> {
-        let tx_size = tx.weight() as usize / 4; // Convert to vsize
+        let tx_size = u64::from(tx.weight()) as usize / 4; // Convert to vsize
         let max_size = self.max_tx_size();
         
         if tx_size > max_size {
@@ -222,7 +222,7 @@ impl MempoolPolicy {
     /// Calculate the minimum fee for a transaction
     pub fn calculate_min_fee(&self, tx: &Transaction) -> u64 {
         let min_relay_fee = self.min_relay_fee();
-        let tx_size = tx.weight() as u64 / 4; // Convert to vsize
+        let tx_size = u64::from(tx.weight()) / 4; // Convert to vsize
         
         // Calculate minimum fee: min_relay_fee * tx_size / 1000
         (min_relay_fee * tx_size + 999) / 1000
